refactor: rename move_generator to move_gen
