+++ src/search.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 #include "board.hpp"
-#include "move_generator.hpp"
+#include "move_gen.hpp"
 #include "search/defs.hpp"
